+++ morebuiltins/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2.3"
+__version__ = "1.2.4"
 __all__ = [
     "morebuiltins.utils",
     "morebuiltins.date",